requirements.txt
openai>=1.0.0
anthropic>=0.5.0
pyyaml>=6.0
nltk>=3.8.1
spacy>=3.5.0
datasketch>=1.5.0
jieba>=0.42.1
matplotlib>=3.6.0
seaborn>=0.12.0
numpy>=1.20.0
pandas>=1.5.0
scikit-learn>=1.0.0
transformers>=4.24.0
pytest>=7.0.0
python-dotenv>=0.20.0
PyMuPDF>=1.22.0 